fix lecturers
tách các component,
sửa nav active ở sidebar.
