+++ client-app/src/features/Lecturers/LecturersRow.tsx
@@ -1,16 +1,10 @@
 import * as React from "react";
 import Collapse from "@mui/material/Collapse";
 import IconButton from "@mui/material/IconButton";
-import Table from "@mui/material/Table";
-import TableBody from "@mui/material/TableBody";
 import TableCell from "@mui/material/TableCell";
-import TableContainer from "@mui/material/TableContainer";
-import TableHead from "@mui/material/TableHead";
 import TableRow from "@mui/material/TableRow";
-import Paper from "@mui/material/Paper";
 import KeyboardArrowDownIcon from "@mui/icons-material/KeyboardArrowDown";
 import KeyboardArrowUpIcon from "@mui/icons-material/KeyboardArrowUp";
-import TablePagination from "@mui/material/TablePagination";
 import List from "@mui/material/List";
 import ListItem from "@mui/material/ListItem";
 import ListItemText from "@mui/material/ListItemText";
@@ -20,6 +14,7 @@ import {Typography} from "@mui/material";
 
 import SimpleBadge from "./SimpaleMail";
 import {AvatarSP} from "../../assets";
+import { observer } from "mobx-react-lite";
 
 function createData(
     lecturers: any,
@@ -42,6 +37,7 @@ function createData(
 }
 
 function Row(props: { row: ReturnType<typeof createData> }) {
+
     const {row} = props;
     const [open, setOpen] = React.useState(false);
 
@@ -127,73 +123,5 @@ function Row(props: { row: ReturnType<typeof createData> }) {
         </React.Fragment>
     );
 }
+export default observer(Row)
 
-const rows = [
-    createData('Nguyễn Thị Phương Thảo', '3/5', 'Mạng Máy Tính', 'Hệ Thống Thông Tin', 'Tiến Sĩ', 'Tiếp Nhận',
-        <SimpleBadge/>),
-    createData('Nguyễn Thị Phương Thảo', '3/5', 'Mạng Máy Tính', 'Hệ Thống Thông Tin', 'Tiến Sĩ', 'Tiếp Nhận',
-        <SimpleBadge/>),
-    createData('Nguyễn Thị Phương Thảo', '3/5', 'Mạng Máy Tính', 'Hệ Thống Thông Tin', 'Tiến Sĩ', 'Tiếp Nhận',
-        <SimpleBadge/>),
-    createData('Nguyễn Thị Phương Thảo', '3/5', 'Mạng Máy Tính', 'Hệ Thống Thông Tin', 'Tiến Sĩ', 'Tiếp Nhận',
-        <SimpleBadge/>),
-    createData('Nguyễn Thị Phương Thảo', '3/5', 'Mạng Máy Tính', 'Hệ Thống Thông Tin', 'Tiến Sĩ', 'Tiếp Nhận',
-        <SimpleBadge/>),
-    createData('Nguyễn Thị Phương Thảo', '3/5', 'Mạng Máy Tính', 'Hệ Thống Thông Tin', 'Tiến Sĩ', 'Tiếp Nhận',
-        <SimpleBadge/>),
-    createData('Nguyễn Thị Phương Thảo', '3/5', 'Mạng Máy Tính', 'Hệ Thống Thông Tin', 'Tiến Sĩ', 'Tiếp Nhận',
-        <SimpleBadge/>),
-];
-
-export default function TableList() {
-    const [page, setPage] = React.useState(0);
-    const [rowsPerPage, setRowsPerPage] = React.useState(5);
-
-    const currentRows = rows.filter((r, ind) => {
-        return ind >= rowsPerPage * page && ind < rowsPerPage * (page + 1);
-    });
-
-    const handleChangePage = (event: unknown, newPage: number) => {
-        setPage(newPage);
-    };
-
-    const handleChangeRowsPerPage = (
-        event: React.ChangeEvent<HTMLInputElement>
-    ) => {
-        setRowsPerPage(parseInt(event.target.value, 10));
-        setPage(0);
-    };
-
-    return (
-        <TableContainer component={Paper} className="table">
-            <Table aria-label="collapsible table">
-                <TableHead>
-                    <TableRow>
-                        <TableCell align="center">Giảng Viên</TableCell>
-                        <TableCell align="center">Đang Hướng Dẫn(SV)</TableCell>
-                        <TableCell align="center">Bộ Môn</TableCell>
-                        <TableCell align="center">Khoa</TableCell>
-                        <TableCell align="center">Học Vị</TableCell>
-                        <TableCell align="center">Trạng thái</TableCell>
-                        <TableCell align="center">Tác vụ</TableCell>
-                        <TableCell/>
-                    </TableRow>
-                </TableHead>
-                <TableBody sx={{background: '#F7F6FE'}}>
-                    {currentRows.map((row, index) => (
-                        <Row key={index} row={row}/>
-                    ))}
-                </TableBody>
-            </Table>
-            <TablePagination
-                rowsPerPageOptions={[5, 10, 25]}
-                component="div"
-                count={rows.length}
-                rowsPerPage={rowsPerPage}
-                page={page}
-                onPageChange={handleChangePage}
-                onRowsPerPageChange={handleChangeRowsPerPage}
-            />
-        </TableContainer>
-    );
-}
