+++ client-app/src/features/Lecturers/lecturers.scss
@@ -43,7 +43,6 @@
           }
 
           .btn_complete {
-            border: 1px solid #787486;
             border-radius: 6px;
           }
         }
